refactor(environment/robot): fold SimThread into bridge; move GymBridge to gym
Dissolve the separate SimThread/run_with_sim surface into RobotBridge._run_on_sim
and serve_bridge, and relocate GymBridge/gym_command/main next to the gym
introspection helpers so each module has one job.
