+++ hud/environment/robot/bridge.py
@@ -1,4 +1,4 @@
-"""Sim-process ``robot`` bridges — and the sim program that serves them.
+"""The sim-process ``robot`` bridge base — and the sim program shape.
 
 The *server* side of the ``robot`` protocol (agent-side client:
 :class:`~hud.capabilities.robot.RobotClient`); both share the wire codec defined
@@ -13,33 +13,32 @@ fans scalar obs back. Single-env is the one-slot degenerate case.
 
 - :class:`RobotBridge` — subclass with your sim: ``reset`` / ``step`` /
   ``get_observation``, and set ``self.contract``. The base owns both channels.
-- :class:`GymBridge` — the generic bridge over any gym-style target (factory,
-  registry id, or a declared env's spec); users get one via ``env.gym(...)``
-  and never subclass it.
-
-This module is also **the sim program** — the one process shape every sim runs
-(the env server never hosts one, it spawns or attaches through an endpoint):
-
-- ``python -m hud.environment.robot.bridge <target>`` — what ``env.gym()``
-  spawns; :func:`gym_command` builds the argv, so both ends of the format live
-  here.
-- :func:`serve_bridge` — the blocking entry a custom sim program calls last.
+  (The generic gym-style bridge, :class:`~.gym.GymBridge`, lives in :mod:`~.gym`.)
+- :func:`serve_bridge` — the blocking entry every sim program calls last.
 
 Inside the sim process the sim owns the main thread and serving runs on a
-background loop; every sim touch routes through the shared
-:class:`~.sim.SimThread`, so bridges stay thread-naive (see :mod:`~.sim`).
+background loop thread; every sim touch is queued to main through
+:meth:`RobotBridge._run_on_sim`. One shape because the hardest sims demand it:
+sims are usually *thread-affine* (every touch must run on the thread that owns
+their GL/device context), and Isaac/Omniverse must own the process main thread
+outright — Kit drives its own main-thread loop and ``env.reset()`` nests
+``run_until_complete``. Cheap CPU envs pay ~nothing: main just blocks on the
+queue.
 """
 
 from __future__ import annotations
 
 import asyncio
 import contextlib
-import inspect
+import queue
 import secrets
+import signal
 import sys
+import threading
 from abc import ABC, abstractmethod
+from collections.abc import Callable
+from concurrent.futures import Future
 from dataclasses import dataclass, field
-from pathlib import Path
 from typing import Any
 
 import numpy as np
@@ -50,17 +49,6 @@ import websockets.exceptions
 # ends of the protocol stay in lockstep (env -> capabilities is the correct direction).
 from hud.capabilities.robot import _packb, _unpackb
 from hud.environment.utils import error, read_frame, reply, send_frame
-from hud.telemetry.robot import to_numpy
-
-from .gym import (
-    action_dim_of,
-    detect_fps,
-    load_or_write_contract,
-    num_envs_of,
-    probe_success,
-    split_observation,
-)
-from .sim import SimThread, run_with_sim
 
 #: Line the sim program prints once its control channel is bound; the spawning
 #: RobotEndpoint reads it from this process's stdout.
@@ -151,8 +139,9 @@ class RobotBridge(ABC):
         #: The env's self-describing wire contract (features, control_rate); the
         #: endpoint fetches it over the control channel to mint the capability.
         self.contract: dict[str, Any] = {}
-        # Every sim touch runs on the process sim thread (see sim.py).
-        self._sim = SimThread.shared()
+        # Sim touches queue to the thread that owns the simulator (see _run_on_sim).
+        self._sim_q: queue.Queue[tuple[Callable[[], Any], Future[Any]]] = queue.Queue()
+        self._sim_ident: int | None = None
         self.num_envs: int = 1
         self._registry = _SlotRegistry()
         self._registry.configure(1)
@@ -164,6 +153,18 @@ class RobotBridge(ABC):
         self.total_reward: float = 0.0
         self.success: bool = False
         self.terminated: bool = False
+
+    async def _run_on_sim(self, fn: Callable[..., Any], *args: Any) -> Any:
+        """Run ``fn(*args)`` on the sim thread, awaited on the caller's loop.
+
+        Before :func:`serve_bridge` binds a sim thread (tests, in-loop use) calls
+        execute inline on the caller — the degenerate single-thread case.
+        """
+        if self._sim_ident is None or threading.get_ident() == self._sim_ident:
+            return fn(*args)  # not serving yet, or already on the sim thread
+        fut: Future[Any] = Future()
+        self._sim_q.put((lambda: fn(*args), fut))
+        return await asyncio.wrap_future(fut)
 
     async def _claim_episode(self, **kwargs: Any) -> dict[str, Any]:
         """Control-plane reset: global sim reset when all free, else claim a free slot."""
@@ -304,6 +305,10 @@ class RobotBridge(ABC):
                 slot.action = None
                 self._action_event.set()  # wake the barrier so it doesn't wait on us
 
+    def _live_slots(self) -> list[_Slot]:
+        """Claimed slots with a live connection — the barrier's participants."""
+        return [s for s in self._registry.claimed() if s.ws is not None]
+
     async def _tick_loop(self) -> None:
         """Gather claimed slots' actions (or timeout → hold), step once, fan out."""
         while True:
@@ -312,7 +317,7 @@ class RobotBridge(ABC):
             except TimeoutError:
                 pass
             self._action_event.clear()
-            claimed = [s for s in self._registry.claimed() if s.ws is not None]
+            claimed = self._live_slots()
             if not claimed:
                 continue
             # Wait until every live claimed slot has an action, or step_timeout elapses.
@@ -329,7 +334,7 @@ class RobotBridge(ABC):
                 self._action_event.clear()
                 with contextlib.suppress(TimeoutError):
                     await asyncio.wait_for(self._action_event.wait(), timeout=remaining)
-                claimed = [s for s in self._registry.claimed() if s.ws is not None]
+                claimed = self._live_slots()
                 if not claimed:
                     break
             if not claimed:
@@ -338,13 +343,11 @@ class RobotBridge(ABC):
             # Stack one row per sim slot (idle/unconnected claimed → hold; free → hold).
             rows = []
             for s in self._registry.slots:
-                if s in claimed and s.action is not None:
-                    rows.append(np.asarray(s.action, dtype=np.float32).reshape(-1))
-                else:
-                    rows.append(np.asarray(hold, dtype=np.float32).reshape(-1))
+                row = s.action if s in claimed and s.action is not None else hold
+                rows.append(np.asarray(row, dtype=np.float32).reshape(-1))
                 s.action = None
             actions = np.stack(rows)
-            await self._sim.call(self.step, actions)
+            await self._run_on_sim(self.step, actions)
             for s in claimed:
                 await self._send_slot_observation(s)
 
@@ -352,7 +355,7 @@ class RobotBridge(ABC):
         """Fan one scalar obs frame to a claimed connection."""
         if slot.ws is None:
             return
-        out = await self._sim.call(self.get_observation)
+        out = await self._run_on_sim(self.get_observation)
         if out is None:
             return
         data, terminated = out
@@ -401,337 +404,76 @@ class RobotBridge(ABC):
         raise ValueError(f"unknown method {method!r}")
 
 
-class GymBridge(RobotBridge):
-    """Serve any gym-style env over the ``robot`` protocol, generically.
-
-    ``target`` is how this process builds the env: a factory callable, or a
-    string — a gymnasium registry id, or a declared env's spec as JSON
-    (:func:`gym_command` reduces an env instance to the latter).
-
-    Task args are partitioned into env-defining **build args** (a change
-    rebuilds the env) and episodic args flowing to ``env.reset(seed=...,
-    options=...)``. For a factory, its signature is the partition — so
-    ``num_envs`` in the signature is the vectorization declaration; for a
-    registry target, ``num_envs`` is the one build arg (``gym.make_vec``).
-    Default build kwargs (e.g. ``num_envs=8`` from ``env.gym(..., num_envs=8)``)
-    fill in when the task does not pass them.
-
-    ``fps`` overrides the detected control rate; ``contract`` is the
-    ``contract.json`` round-trip path (load beats derive, so user edits stick;
-    ``None`` disables the file).
-    """
-
-    def __init__(
-        self,
-        target: Any,
-        *,
-        fps: int | None = None,
-        contract: str | Path | None = "contract.json",
-        **defaults: Any,
-    ) -> None:
-        host = defaults.pop("host", "127.0.0.1")
-        port = int(defaults.pop("port", 0))
-        super().__init__(host=host, port=port)
-        self._target = target
-        self._fps = fps
-        self._contract_path = contract
-        self._defaults = defaults  # e.g. num_envs from env.gym(..., num_envs=8)
-        # Task args that define the env build (everything else is episodic).
-        if callable(target):
-            self._build_params = {
-                n
-                for n, p in inspect.signature(target).parameters.items()
-                if p.kind in (p.POSITIONAL_OR_KEYWORD, p.KEYWORD_ONLY)
-            }
-        else:
-            self._build_params = {"num_envs"}  # registry targets: vectorization only
-        self.env: Any = None
-        self.batched = False  # env carries a leading [N] dim (any env exposing num_envs)
-        self._obs: Any = None  # latest observation (reset or step)
-        self._instance: Any = None  # current env-defining args; a mismatch rebuilds
-        self._is_torch = False
-        # Per-slot episode scoring, sticky until the next reset.
-        self._done: np.ndarray = np.zeros(1, dtype=bool)
-        self._success: np.ndarray = np.zeros(1, dtype=bool)
-        self._acc_reward: np.ndarray = np.zeros(1)
-        self._step_reward: np.ndarray = np.zeros(1)  # last env.step reward (RL wire sibling)
-        self._seen_success = False  # any env-reported success signal this episode
-
-    # ── env lifecycle (all sim touches on the sim thread) ───────────────────────
-
-    async def ensure_env(self, **task_args: Any) -> None:
-        """Build the env (factory defaults unless task args say otherwise). Idempotent."""
-        if self.env is None:
-            await self._sim.call(self._sync_reset, task_args)
-
-    async def start(self) -> None:
-        """Build the env and derive/load the contract, then bring up the wire."""
-        await self.ensure_env()
-        state, frames = self.sample_observation()
-        existed = self._contract_path is not None and Path(self._contract_path).exists()
-        self.contract = load_or_write_contract(
-            self._contract_path,
-            state,
-            frames,
-            action_dim_of(self.env, batched=self.batched),
-            self._fps or detect_fps(self.env),
-        )
-        if not existed and self._contract_path is not None:
-            print(f"[env] wrote {self._contract_path} (edit names to relabel plots)", flush=True)
-        await super().start()
-
-    async def reset(self, **task_args: Any) -> str:
-        return await self._sim.call(self._sync_reset, task_args)
-
-    async def stop(self) -> None:
-        await super().stop()
-        if self.env is not None:
-            await self._sim.call(self.env.close)
-            self.env = None
-
-    def _sync_reset(self, task_args: dict[str, Any]) -> str:
-        # Defaults from env.gym(..., num_envs=N) fill in when the task omits them.
-        merged = {**self._defaults, **task_args}
-        build = {k: v for k, v in merged.items() if k in self._build_params}
-        episodic = {k: v for k, v in merged.items() if k not in self._build_params}
-        key = tuple(sorted(build.items()))
-        if self.env is None or key != self._instance:
-            if self.env is not None:
-                self.env.close()
-            self.env = self._build_env(build)
-            self._instance = key
-            n = num_envs_of(self.env)
-            self.batched = n is not None
-            self.num_envs = n or 1
-        seed = episodic.pop("seed", None)
-        obs, _ = self.env.reset(seed=seed, options=episodic or None)
-        self._obs = obs  # the first frame an agent sees on connect/reset
-        self._is_torch = "torch" in type(_first_leaf(obs)).__module__
-        self._done = np.zeros(self.num_envs, dtype=bool)
-        self._success = np.zeros(self.num_envs, dtype=bool)
-        self._acc_reward = np.zeros(self.num_envs)
-        self._step_reward = np.zeros(self.num_envs, dtype=np.float32)
-        self._seen_success = False
-        return self._prompt(merged)
-
-    def _build_env(self, build: dict[str, Any]) -> Any:
-        """Build the env from the target: factory call, or registry make/make_vec."""
-        if callable(self._target):
-            return self._target(**build)
-        import gymnasium
-        from gymnasium.envs.registration import EnvSpec
-
-        env_id, kwargs = self._target, dict(build)
-        if env_id.startswith("{"):  # a declared env's spec, re-made in this process
-            spec = EnvSpec.from_json(env_id)
-            env_id, kwargs = spec.id, {**spec.kwargs, **kwargs}
-        if "num_envs" in kwargs:
-            return gymnasium.make_vec(env_id, **kwargs)
-        return gymnasium.make(env_id, **kwargs)
-
-    def _prompt(self, task_args: dict[str, Any]) -> str:
-        base = getattr(self.env, "unwrapped", self.env)
-        for attr in ("task_description", "instruction"):
-            text = getattr(getattr(base, "cfg", None), attr, None) or getattr(base, attr, None)
-            if isinstance(text, str) and text:
-                return text
-        return ", ".join(f"{k}={v}" for k, v in sorted(task_args.items())) or "run the task"
-
-    def sample_observation(self) -> tuple[dict[str, Any], dict[str, Any]]:
-        """One per-env ``(state, frames)`` sample for contract derivation (post-build)."""
-        state, frames = split_observation(self._obs, batched=self.batched)
-        if self.batched:
-            state = {k: to_numpy(v)[0] for k, v in state.items()}
-            frames = {k: to_numpy(v)[0] for k, v in frames.items()}
-        return state, frames
-
-    # ── bridge protocol ──────────────────────────────────────────────────────────
-
-    def step(self, action: np.ndarray) -> None:
-        act: Any = np.array(action, dtype=np.float32)  # wire buffer is read-only
-        if not self.batched:
-            act = act[0] if act.ndim > 1 else act  # single plain env: drop the batch dim
-        elif act.ndim == 1:
-            act = act[None]
-        # The wire carries floats; discrete/int action spaces need their dtype + shape back.
-        space = getattr(self.env, "action_space", None)
-        dtype = getattr(space, "dtype", None)
-        if dtype is not None and np.issubdtype(dtype, np.integer):
-            act = act.astype(dtype).reshape(getattr(space, "shape", act.shape) or ())
-        if self._is_torch:
-            import torch
-
-            base = getattr(self.env, "unwrapped", self.env)
-            act = torch.as_tensor(act, device=getattr(base, "device", None))
-        obs, reward, terminated, truncated, info = self.env.step(act)
-        self._obs = obs
-        done = np.atleast_1d(to_numpy(terminated)).astype(bool) | np.atleast_1d(
-            to_numpy(truncated)
-        ).astype(bool)
-        # Per-step reward for the RL wire (get_observation attaches it as a sibling).
-        self._step_reward = np.atleast_1d(to_numpy(reward)).astype(np.float32)
-        self._acc_reward += self._step_reward * ~self._done
-        newly = done & ~self._done
-        if newly.any():
-            success = self._resolve_success(info)
-            if success is not None:
-                self._seen_success = True
-                self._success |= success & newly
-        self._done |= done
-
-    def _resolve_success(self, info: Any) -> np.ndarray | None:
-        """Env-reported success at a done step: info keys, else Isaac's termination term."""
-        found = probe_success(info, num_envs=self.num_envs)
-        if found is not None:
-            return found
-        base = getattr(self.env, "unwrapped", self.env)
-        manager = getattr(base, "termination_manager", None)
-        if manager is not None:
-            try:
-                return np.atleast_1d(to_numpy(manager.get_term("success"))).astype(bool)
-            except Exception:
-                return None
-        return None
-
-    def get_observation(self) -> tuple[dict[str, np.ndarray], np.ndarray] | None:
-        """Always ``[N, ...]`` arrays + ``[N]`` terminated for the barrier fan-out."""
-        if self.env is None or self._obs is None:
-            return None
-        state, frames = split_observation(self._obs, batched=self.batched)
-        data = {k: to_numpy(v) for k, v in {**state, **frames}.items()}
-        # Last env.step reward rides along for RL collection (client lifts it
-        # out of "data" to a top-level sibling, like "terminated").
-        data["reward"] = self._step_reward
-        if not self.batched:
-            # Plain single env: lift scalars to a batch of one for uniform slicing.
-            data = {k: (v if getattr(v, "ndim", 0) >= 1 and v.shape[:1] == (1,) else np.asarray(v)[None])
-                    for k, v in data.items()}
-        return data, np.asarray(self._done, dtype=bool).reshape(self.num_envs)
-
-    def result_slots(self) -> list[dict[str, Any]]:
-        """Per-slot grades: env-reported success when available, else accumulated reward."""
-        # The env's own success check outranks accumulated shaped reward.
-        scores = self._success if self._seen_success else self._acc_reward
-        return [
-            {
-                "score": float(scores[i]),
-                "success": bool(self._success[i]),
-                "total_reward": float(self._acc_reward[i]),
-            }
-            for i in range(self.num_envs)
-        ]
-
-
-def _first_leaf(obs: Any) -> Any:
-    while isinstance(obs, dict):
-        obs = next(iter(obs.values()))
-    return obs
-
-
-# ── the sim program (what env.gym() spawns; custom sims call serve_bridge) ────
+# ── the sim program shape (custom sims call serve_bridge last) ────────────────
 
 
 def serve_bridge(bridge: RobotBridge, *, host: str = "127.0.0.1", port: int = 0) -> None:
     """Serve *bridge*, blocking for the process's lifetime — the sim program's last call.
 
-    Brings up the robot WebSocket and the control side channel (announcing its
-    port on stdout as ``HUD_SIM_PORT=<port>``), with the sim owning the main
-    thread. SIGTERM / Ctrl-C tear the bridge down on the way out.
+    Serving — the robot WebSocket and the control side channel (port announced
+    on stdout as ``HUD_SIM_PORT=<port>``) — runs on a background loop thread;
+    the calling (main) thread becomes the sim thread, executing every touch the
+    bridge queues (see :meth:`RobotBridge._run_on_sim`). SIGTERM / Ctrl-C cancel
+    serving; the sim keeps draining through teardown (``stop()`` touches it too).
     """
+    # Claim main as the sim thread before serving starts, so no touch slips
+    # through inline on the wrong thread.
+    bridge._sim_ident = threading.get_ident()
+    loop = asyncio.new_event_loop()
+    done = threading.Event()
+    serve_task: asyncio.Task[Any] | None = None
 
     async def _serve() -> None:
+        nonlocal serve_task
+        serve_task = asyncio.current_task()
         await bridge.start()  # WS bound first, so `url` is concrete before any control call
         server = await bridge.serve_control(host, port)
-        bound = server.sockets[0].getsockname()[1]
-        print(f"{PORT_ANNOUNCEMENT}{bound}", flush=True)
+        print(f"{PORT_ANNOUNCEMENT}{server.sockets[0].getsockname()[1]}", flush=True)
         try:
             await asyncio.Event().wait()  # until SIGTERM / Ctrl-C cancels
         finally:
             server.close()
             await bridge.stop()
 
-    run_with_sim(_serve)
+    def _thread() -> None:
+        asyncio.set_event_loop(loop)
+        try:
+            with contextlib.suppress(asyncio.CancelledError):
+                loop.run_until_complete(_serve())
+        finally:
+            done.set()
+
+    def _cancel(*_: object) -> None:
+        if serve_task is not None:
+            loop.call_soon_threadsafe(serve_task.cancel)
+
+    with contextlib.suppress(ValueError):  # signals are main-thread-only; tests may not be
+        signal.signal(signal.SIGTERM, _cancel)
+    thread = threading.Thread(target=_thread, name="hud-serve", daemon=True)
+    thread.start()
+
+    # The sim loop: execute queued touches until serving exits, pumping Kit when
+    # Omniverse is loaded (it needs continuous main-thread updates).
+    while not done.is_set():
+        try:
+            # Re-checked every pass: Kit may finish loading after serving starts.
+            kit = sys.modules.get("omni.kit.app")
+            touches: list[tuple[Callable[[], Any], Future[Any]]] = []
+            with contextlib.suppress(queue.Empty):
+                touches.append(bridge._sim_q.get(timeout=0.002 if kit else 0.05))
+                while True:  # drain the backlog behind the first touch
+                    touches.append(bridge._sim_q.get_nowait())
+            for fn, fut in touches:
+                if fut.set_running_or_notify_cancel():
+                    try:
+                        fut.set_result(fn())
+                    except BaseException as exc:  # propagate to the awaiting caller
+                        fut.set_exception(exc)
+            if kit:
+                kit.get_app().update()
+        except KeyboardInterrupt:  # Ctrl-C: cancel serving, keep draining through teardown
+            _cancel()
+    thread.join(timeout=30)
 
 
-def gym_command(
-    target: Any,
-    *,
-    fps: int | None = None,
-    contract: str | None = "contract.json",
-    **defaults: Any,
-) -> list[str]:
-    """Build the command line that spawns *target* as a sim process (what ``env.gym`` runs).
-
-    ``env.py`` only declares the sim, it never runs it — this builds the argv
-    for the child process that will. A live env object can't cross that
-    boundary, only text can, so *target* is reduced to something the child can
-    rebuild itself: a factory's source path, a registry id as-is, or a
-    constructed env's spec (id + kwargs) — closed here since only the spec
-    crosses over; the child calls ``gym.make`` again for its own instance.
-    ``fps`` / ``contract`` / build defaults (e.g. ``num_envs=``) flow through
-    to the :class:`GymBridge` the CLI builds.
-    """
-    if callable(target):
-        name = getattr(target, "__qualname__", "")
-        if not name or "." in name or "<" in name:
-            raise ValueError(f"env.gym factory must be a module-level callable, got {target!r}")
-        target = f"{inspect.getfile(target)}:{name}"
-    elif not isinstance(target, str):
-        spec = getattr(target, "spec", None)  # registry-made envs carry their EnvSpec
-        if spec is None:
-            raise ValueError(
-                f"env.gym: {target!r} has no registry spec; pass a gymnasium id "
-                "or a module-level factory instead"
-            )
-        target_env, target = target, spec.to_json()
-        target_env.close()  # only the spec crosses; free the declaration-time instance
-    cmd = [sys.executable, "-m", "hud.environment.robot.bridge", target]
-    if fps is not None:
-        cmd += ["--fps", str(fps)]
-    if contract is not None:
-        cmd += ["--contract", str(contract)]
-    # Build defaults (num_envs, etc.) as --key value pairs the CLI re-applies.
-    for key, value in defaults.items():
-        cmd += [f"--{key.replace('_', '-')}", str(value)]
-    return cmd
-
-
-def main() -> None:
-    import argparse
-
-    from hud.utils.modules import load_module
-
-    parser = argparse.ArgumentParser(description="Serve a gym-style env as a robot sim.")
-    parser.add_argument("target", help="'path/to/module.py:factory', a registry id, or spec JSON.")
-    parser.add_argument("--fps", type=int, default=None, help="Control-rate override.")
-    parser.add_argument("--contract", default=None, help="contract.json round-trip path.")
-    parser.add_argument("--host", default="127.0.0.1", help="Control-channel interface.")
-    parser.add_argument("--port", type=int, default=0, help="Control-channel port (0 = ephemeral).")
-    parser.add_argument("--num-envs", type=int, default=None, help="Vectorized slot count.")
-    args, unknown = parser.parse_known_args()
-
-    target: Any = args.target
-    if ".py:" in target:  # factory by source path; ids / spec JSON pass through
-        path, _, attr = target.rpartition(":")
-        target = getattr(load_module(path), attr)
-    defaults: dict[str, Any] = {}
-    if args.num_envs is not None:
-        defaults["num_envs"] = args.num_envs
-    # Accept further --key value pairs as build defaults.
-    i = 0
-    while i < len(unknown):
-        flag = unknown[i]
-        if flag.startswith("--") and i + 1 < len(unknown):
-            defaults[flag[2:].replace("-", "_")] = unknown[i + 1]
-            i += 2
-        else:
-            i += 1
-    bridge = GymBridge(target, fps=args.fps, contract=args.contract, **defaults)
-    serve_bridge(bridge, host=args.host, port=args.port)
-
-
-if __name__ == "__main__":
-    main()
-
-
-__all__ = ["PORT_ANNOUNCEMENT", "GymBridge", "RobotBridge", "gym_command", "serve_bridge"]
+__all__ = ["PORT_ANNOUNCEMENT", "RobotBridge", "serve_bridge"]
